fix: moved imports to fix hot reload
