+++ frontend/content/checkboxOptions.ts
@@ -1,4 +1,6 @@
-import { ComponentOptions } from '../src/types/form'
+import { heirsTypes } from '../src/app/(dynamic)/last-will/editor/heirs/page'
+import { HeirsTypes } from '../src/store/last-will/heirs/state'
+import { ComponentOptions, DropdownButtonOptions } from '../src/types/form'
 
 export const testatorMoreInfosOptions: ComponentOptions[] = [
 	{
@@ -36,3 +38,11 @@ export const personMoreInfosOptions: ComponentOptions[] = [
 		label: 'Ist die Person insolvent?',
 	},
 ]
+
+export const getPersonAddHeirsOptions = (setDropdownOption: SetDropdownOptionFunction): DropdownButtonOptions[] =>
+	Object.entries(heirsTypes).map(([type, label]) => ({
+		onClick: () => setDropdownOption(type as HeirsTypes),
+		label: `${label} hinzufügen`,
+	}))
+
+export type SetDropdownOptionFunction = (type: HeirsTypes) => void
